requirements.txt
lpips==0.1.4
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.5.0
torch==1.10.2+cu113
torchvision==0.11.3+cu113
